Specify precision in python code

+++ scripts/matrix_wrappers.py
@@ -30,7 +30,7 @@ class MMatrix(MatrixWrapper):
         return self.matvec(x)
 
     def solve(self, x):
-        x = x.astype(np.float32)
+        # x = x.astype(np.float32)
         return self.LU.solve(x)
 
     def materialize_inverse(self):
